\name{mbiemaps-package}
\alias{mbiemaps-package}
\alias{mbiemaps}
\docType{package}
\title{
Maps commonly used in MBIE
}
\description{
Spatial Polygon Data Frames of maps commonly used in MBIE
}
\details{
\tabular{ll}{
Package: \tab mbiemaps\cr
Type: \tab Package\cr
Version: \tab 1.2.3\cr
Date: \tab 2016-3-23\cr
License: \tab (check)
}
}
\author{
Peter Ellis

Maintainer: Eric Wu <Eric.Wu@mbie.govt.nz> & Sam Vennell <Sam.Vennell@mbie.govt.nz>

}
\references{
All the original shapefiles were sourced from \href{http://www.stats.govt.nz/browse_for_stats/people_and_communities/Geographic-areas/digital-boundary-files.aspx}{http://www.stats.govt.nz/browse_for_stats/people_and_communities/Geographic-areas/digital-boundary-files.aspx}, where they are available under a Creative Commons license
}

\keyword{ package }
\seealso{
}
\examples{
}
