% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbiemaps-package.R
\docType{package}
\name{mbiemaps-package}
\alias{mbiemaps}
\alias{mbiemaps-package}
\title{mbiemaps: Maps commonly used in MBIE}
\description{
The mbiemaps package provides Spatial Polygon Data Frames of maps commonly used in MBIE
}
\details{
{   \\tabular{ll}{ Package: \\tab mbiemaps\\cr Type: \\tab Package\\cr Version: \\tab 1.2.4\\cr Date: \\tab 2016-7-05\\cr License: \\tab (check)} }
}
\author{
: Peter Ellis <Peter.Ellis@mbie.govt.nz>; Eric Wu <Eric.Wu@mbie.govt.nz>; Sam Vennell <Sam.Vennell@mbie.govt.nz>
}
\references{
All the original shapefiles were sourced from \url{http://www.stats.govt.nz/browse_for_stats/people_and_communities/Geographic-areas/digital-boundary-files.aspx}, where they are available under a Creative Commons license
}
\keyword{package}

